requirements.txt
boto3>=1.3.1
pyyaml
